Move console classes to their own module
